Update overview.md
It is important to refer to "convenience stores" as Deliveroo reviewed the description we made of them on the Apps page and insisted on the change. 
I also think we should include the Integration Feature section here.
Prerequisites section go last. I think we should keep this section in the Overview page, and not remove it as we had discussed with Antoine at some point.