update-deliveroo-configuration.md
Please note that we changed the termonology. Point of Sales should always be referred to as EPOS (more modern) and not POS.

Regarding **Payment Methods**, I don't understand it.... 
It is the Deliveroo reference they need to input here? 
What about the EPOS ref code? Some of them need a reference.

Same for the **Discount** section. I guess the name and ref codes should be found in the EPOS, correct?

We need to include something along the lines of "Refer to your connected EPOS documentation on the HubRise website to verify" every time the user needs to check if an EPOS ref code is needed. This will at least explain to the reader that the reference can be fond in the EPOS.

FYI: The more Uber Eats connections I do, the more I think we will need a specific section for every EPOS documentation explaining how things are done. I will make a note of this in the iKento Board.