update-ubereats-overview.md
I have not made changes to the content of this page as they are the same comments I have done for Deliveroo. I trust you guys (@antmonnier and/or @atorin) will do this, when it's relevant :-)

Just one comment. I think we should not use refer to the Uber **ecosystem** as this is a term we use for HubRise and our ecosystem is very different from theirs.